OctopusCore/requirements.txt
PyYAML
Flask-Cors
sendgrid
PyJWT
requests
passlib
requests-toolbelt
ecdsa